refactor(project create): split RunE into phase-specific files
project_create.go was 888 lines with a ~570-line RunE closure mixing
flag parsing, interactive form, validation, scaffolding, composer
install, and summary rendering. Split into one orchestrator plus four
phase files in the same package:

- project_create.go        — cobra command + RunE orchestrator,
                             createOptions, parseCreateFlags,
                             applyNonInteractiveDefaults, init
- project_create_form.go   — runCreateForm (interactive huh form loop)
- project_create_validate.go — validateAndPreflight, security advisory
                               and incompatibility prompts, version
                               resolution helpers
- project_create_scaffold.go — scaffoldProject + setupDeployment +
                               setupCI, embeds
- project_create_install.go  — installAndFinalize, printCreateSummary,
                               runComposerInstall

No behavior change. Existing TestResolveVersion / TestSetupDeployment /
TestSetupCI / TestValidDeploymentMethods / TestValidCISystems still
pass against the new layout.

Refs CODE_IMPROVEMENTS item #3.
